pyproject.toml
[project]
name = "ai-health-example"
version = "0.1.0"
description = "AI 헬스케어 파이널 프로젝트 예시용"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "redis>=7.1.0",
    "tomlkit>=0.14.0",
]

[dependency-groups]
ai = [
    "asyncmy>=0.2.11",
    "celery>=5.4.0",
    "redis>=7.1.0",
    "tortoise-orm>=0.25.3",
    "scikit-learn>=1.8.0",
    "torch",
    "torchvision",
    "torchaudio",
    "sentence-transformers>=5.2.0",
]
app = [
    "aerich>=0.9.2",
    "asyncmy>=0.2.11",
    "bcrypt<=4.0.1",
    "celery>=5.4.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "orjson>=3.11.5",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "tortoise-orm>=0.25.3",
    "uvicorn>=0.40.0",
]
dev = [
    "coverage>=7.13.2",
    "mypy>=1.19.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
    "types-passlib>=1.7.7.20250602",
    "types-python-dateutil>=2.9.0.20260124",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.aerich]
tortoise_orm = "app.core.db.databases.TORTOISE_ORM"
location = "./app/core/db/migrations"
src_folder = "./."

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90", "B", "UP", "N"]
ignore = ["UP046", "E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"db/migrations/*" = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"